+++ modules/geo/geo.ts
@@ -1,32 +1,34 @@
+import type { Vec2 } from './vector';
+
 // constants
 var TAU = 2 * Math.PI;
 const EQUATORIAL_RADIUS = 6.3781370E6;
 const POLAR_RADIUS = 6.3567523E6;
 
 
-export function geoLatToMeters(dLat) {
+export function geoLatToMeters(dLat: number) {
     return dLat * (TAU * POLAR_RADIUS / 360);
 }
 
 
-export function geoLonToMeters(dLon, atLat) {
+export function geoLonToMeters(dLon: number, atLat: number) {
     return Math.abs(atLat) >= 90 ? 0 :
         dLon * (TAU * EQUATORIAL_RADIUS / 360) * Math.abs(Math.cos(atLat * (Math.PI / 180)));
 }
 
 
-export function geoMetersToLat(m) {
+export function geoMetersToLat(m: number) {
     return m / (TAU * POLAR_RADIUS / 360);
 }
 
 
-export function geoMetersToLon(m, atLat) {
+export function geoMetersToLon(m: number, atLat: number) {
     return Math.abs(atLat) >= 90 ? 0 :
         m / (TAU * EQUATORIAL_RADIUS / 360) / Math.abs(Math.cos(atLat * (Math.PI / 180)));
 }
 
 
-export function geoMetersToOffset(meters, tileSize) {
+export function geoMetersToOffset(meters: Vec2, tileSize?: number): Vec2 {
     tileSize = tileSize || 256;
     return [
         meters[0] * tileSize / (TAU * EQUATORIAL_RADIUS),
@@ -35,7 +37,7 @@ export function geoMetersToOffset(meters, tileSize) {
 }
 
 
-export function geoOffsetToMeters(offset, tileSize) {
+export function geoOffsetToMeters(offset: Vec2, tileSize?: number): Vec2 {
     tileSize = tileSize || 256;
     return [
         offset[0] * TAU * EQUATORIAL_RADIUS / tileSize,
@@ -45,7 +47,7 @@ export function geoOffsetToMeters(offset, tileSize) {
 
 
 // Equirectangular approximation of spherical distances on Earth
-export function geoSphericalDistance(a, b) {
+export function geoSphericalDistance(a: Vec2, b: Vec2) {
     var x = geoLonToMeters(a[0] - b[0], (a[1] + b[1]) / 2);
     var y = geoLatToMeters(a[1] - b[1]);
     return Math.sqrt((x * x) + (y * y));
@@ -53,7 +55,7 @@ export function geoSphericalDistance(a, b) {
 
 
 // scale to zoom
-export function geoScaleToZoom(k, tileSize) {
+export function geoScaleToZoom(k: number, tileSize?: number) {
     tileSize = tileSize || 256;
     var log2ts = Math.log(tileSize) * Math.LOG2E;
     return Math.log(k * TAU) / Math.LN2 - log2ts;
@@ -61,14 +63,14 @@ export function geoScaleToZoom(k, tileSize) {
 
 
 // zoom to scale
-export function geoZoomToScale(z, tileSize) {
+export function geoZoomToScale(z: number, tileSize?: number) {
     tileSize = tileSize || 256;
     return tileSize * Math.pow(2, z) / TAU;
 }
 
 
 // returns info about the node from `nodes` closest to the given `point`
-export function geoSphericalClosestNode(nodes, point) {
+export function geoSphericalClosestNode(nodes: iD.OsmNode[], point: Vec2) {
     var minDistance = Infinity, distance;
     var indexOfMin;
 
@@ -76,7 +78,7 @@ export function geoSphericalClosestNode(nodes, point) {
         distance = geoSphericalDistance(nodes[i].loc, point);
         if (distance < minDistance) {
             minDistance = distance;
-            indexOfMin = i;
+            indexOfMin = +i;
         }
     }