+++ modules/geo/vector.ts
@@ -1,5 +1,7 @@
+export type Vec2 = [x: number, y: number];
+
 // vector equals
-export function geoVecEqual(a, b, epsilon) {
+export function geoVecEqual(a: Vec2, b: Vec2, epsilon?: number) {
     if (epsilon) {
         return (Math.abs(a[0] - b[0]) <= epsilon) && (Math.abs(a[1] - b[1]) <= epsilon);
     } else {
@@ -8,27 +10,27 @@ export function geoVecEqual(a, b, epsilon) {
 }
 
 // vector addition
-export function geoVecAdd(a, b) {
+export function geoVecAdd(a: Vec2, b: Vec2): Vec2 {
     return [ a[0] + b[0], a[1] + b[1] ];
 }
 
 // vector subtraction
-export function geoVecSubtract(a, b) {
+export function geoVecSubtract(a: Vec2, b: Vec2): Vec2 {
     return [ a[0] - b[0], a[1] - b[1] ];
 }
 
 // vector scaling
-export function geoVecScale(a, mag) {
+export function geoVecScale(a: Vec2, mag: number): Vec2 {
     return [ a[0] * mag, a[1] * mag ];
 }
 
 // vector rounding (was: geoRoundCoordinates)
-export function geoVecFloor(a) {
+export function geoVecFloor(a: Vec2): Vec2 {
     return [ Math.floor(a[0]), Math.floor(a[1]) ];
 }
 
 // linear interpolation
-export function geoVecInterp(a, b, t) {
+export function geoVecInterp(a: Vec2, b: Vec2, t: number): Vec2 {
     return [
         a[0] + (b[0] - a[0]) * t,
         a[1] + (b[1] - a[1]) * t
@@ -36,12 +38,12 @@ export function geoVecInterp(a, b, t) {
 }
 
 // http://jsperf.com/id-dist-optimization
-export function geoVecLength(a, b) {
+export function geoVecLength(a: Vec2, b: Vec2) {
     return Math.sqrt(geoVecLengthSquare(a,b));
 }
 
 // length of vector raised to the power two
-export function geoVecLengthSquare(a, b) {
+export function geoVecLengthSquare(a: Vec2, b?: Vec2) {
     b = b || [0, 0];
     var x = a[0] - b[0];
     var y = a[1] - b[1];
@@ -49,7 +51,7 @@ export function geoVecLengthSquare(a, b) {
 }
 
 // get a unit vector
-export function geoVecNormalize(a) {
+export function geoVecNormalize(a: Vec2): Vec2 {
     var length = Math.sqrt((a[0] * a[0]) + (a[1] * a[1]));
     if (length !== 0) {
         return geoVecScale(a, 1 / length);
@@ -59,12 +61,12 @@ export function geoVecNormalize(a) {
 
 // Return the counterclockwise angle in the range (-pi, pi)
 // between the positive X axis and the line intersecting a and b.
-export function geoVecAngle(a, b) {
+export function geoVecAngle(a: Vec2, b: Vec2) {
     return Math.atan2(b[1] - a[1], b[0] - a[0]);
 }
 
 // dot product
-export function geoVecDot(a, b, origin) {
+export function geoVecDot(a: Vec2, b: Vec2, origin?: Vec2) {
     origin = origin || [0, 0];
     var p = geoVecSubtract(a, origin);
     var q = geoVecSubtract(b, origin);
@@ -72,7 +74,7 @@ export function geoVecDot(a, b, origin) {
 }
 
 // normalized dot product
-export function geoVecNormalizedDot(a, b, origin) {
+export function geoVecNormalizedDot(a: Vec2, b: Vec2, origin?: Vec2) {
     origin = origin || [0, 0];
     var p = geoVecNormalize(geoVecSubtract(a, origin));
     var q = geoVecNormalize(geoVecSubtract(b, origin));
@@ -82,7 +84,7 @@ export function geoVecNormalizedDot(a, b, origin) {
 // 2D cross product of OA and OB vectors, returns magnitude of Z vector
 // Returns a positive value, if OAB makes a counter-clockwise turn,
 // negative for clockwise turn, and zero if the points are collinear.
-export function geoVecCross(a, b, origin) {
+export function geoVecCross(a: Vec2, b: Vec2, origin?: Vec2) {
     origin = origin || [0, 0];
     var p = geoVecSubtract(a, origin);
     var q = geoVecSubtract(b, origin);
@@ -91,7 +93,7 @@ export function geoVecCross(a, b, origin) {
 
 
 // find closest orthogonal projection of point onto points array
-export function geoVecProject(a, points) {
+export function geoVecProject(a: Vec2, points: Vec2[]) {
     var min = Infinity;
     var idx;
     var target;
@@ -101,7 +103,7 @@ export function geoVecProject(a, points) {
         var s = geoVecSubtract(points[i + 1], o);
         var v = geoVecSubtract(a, o);
         var proj = geoVecDot(v, s) / geoVecDot(s, s);
-        var p;
+        var p: Vec2;
 
         if (proj < 0) {
             p = o;