+++ modules/ui/fields/input.js
@@ -111,7 +111,7 @@ export function uiFieldText(field, context) {
         if (field.type === 'tel') {
             updatePhonePlaceholder();
 
-        } else if (field.type === 'number') {
+        } else if (field.type === 'number' || field.type === 'integer') {
             var rtl = (localizer.textDirection() === 'rtl');
 
             input.attr('type', 'text');
@@ -441,7 +441,7 @@ export function uiFieldText(field, context) {
             if (!val && getVals(_tags).size > 1) return;
 
             let displayVal = val;
-            if (field.type === 'number' && val) {
+            if ((field.type === 'number' || field.type === 'integer') && val) {
                 const numbers = val.split(';').map(v => {
                     if (likelyRawNumberFormat.test(v)) {
                         // input number likely in "raw" format
@@ -506,7 +506,7 @@ export function uiFieldText(field, context) {
         var val = vals.size === 1 ? [...vals][0] ?? '' : '';
         var shouldUpdate;
 
-        if (field.type === 'number' && val) {
+        if ((field.type === 'number' || field.type === 'integer') && val) {
             var numbers = val.split(';');
             var oriNumbers = utilGetSetValue(input).split(';');
             if (numbers.length !== oriNumbers.length) shouldUpdate = true;
@@ -548,7 +548,7 @@ export function uiFieldText(field, context) {
             .attr('placeholder', isMixed ? t('inspector.multiple_values') : (field.placeholder() || t('inspector.unknown')))
             .classed('mixed', isMixed);
 
-        if (field.type === 'number') {
+        if (field.type === 'number' || field.type === 'integer') {
             const buttons = wrap.selectAll('.increment, .decrement');
             if (isMixed) {
                 buttons.attr('disabled', 'disabled').classed('disabled', true);